startercar
StarterCar is an encyclopedia of great first choices for the young car
enthusiast. Search the database or simply click through the entries to
find inspiration for your first car! Utilizes ajax to make asynchronous
calls to PHP scripts, which query a MySQL database to run the
autocompleting search bar and load results. Bootstrap ensures the
front-end looks good in any resolution.